Looked at the parts Sam just added.
I commented on a couple of things, but the important change came in line 47. Added "$" to the last "UserID" and removed single quotes from around the curly braces.

I think that the extra query for the UserID was redundant (that should just be the value of $result), but I'm not going to change that since you put the time in to do it, and it doesn't hurt to leave there.