Finished queries
Each query now does what you need it to do. The TF and MC both return a number (as explained in the comments), and the SA returns a list of strings (the responses themselves). I didn't write up the PHP necessary to extract this information, but it shouldn't be very much.